+++ src/Timestepping/Timestepping.jl
@@ -44,6 +44,7 @@ using ..BiotSavart:
 # See https://docs.sciml.ai/CommonSolve/stable/
 import CommonSolve: init, solve!, step!
 
+using ForwardDiff: ForwardDiff
 using TimerOutputs: TimerOutputs, TimerOutput, @timeit, reset_timer!
 
 abstract type AbstractProblem end
@@ -147,6 +148,7 @@ struct VortexFilamentSolver{
         FastTerm <: FastBiotSavartTerm,
         Timer <: TimerOutput,
         Callback <: Function,
+        ExternalForcing <: NamedTuple,
         AdvectFunction <: Function,
         RHSFunction <: Function,
     } <: AbstractSolver
@@ -163,6 +165,7 @@ struct VortexFilamentSolver{
     cache_timestepper :: CacheTimestepper
     fast_term         :: FastTerm
     callback :: Callback
+    external_forcing :: ExternalForcing  # velocity and streamfunction forcing
     to       :: Timer
     advect!  :: AdvectFunction  # function for advecting filaments with a known velocity
     rhs!     :: RHSFunction     # function for estimating filament velocities (and sometimes streamfunction) from their positions
@@ -186,6 +189,9 @@ function Base.show(io::IO, iter::VortexFilamentSolver)
     summary(io, iter.cache_timestepper)
     print(io, "\n - `fast_term`: ", iter.fast_term)
     print(io, "\n - `callback`: Function (`", _printable_function(iter.callback), "`)")
+    for (name, func) ∈ pairs(iter.external_forcing)
+        func === nothing || print(io, "\n - `external_forcing.$name`: Function (`$func`)")
+    end
     print(io, "\n - `advect!`: Function")
     print(io, "\n - `rhs!`: Function")
     print(io, "\n - `to`: ")
@@ -212,7 +218,7 @@ _printable_function(f::TimerOutputs.InstrumentedFunction) = f.func
 get_dt(iter::VortexFilamentSolver) = iter.time.dt
 get_t(iter::VortexFilamentSolver) = iter.time.t
 
-"""
+@doc raw"""
     init(prob::VortexFilamentProblem, scheme::TemporalScheme; dt::Real, kws...) -> VortexFilamentSolver
 
 Initialise vortex filament problem.
@@ -250,6 +256,70 @@ either [`step!`](@ref) or [`solve!`](@ref).
 
 - `timer = TimerOutput("VortexFilament")`: an optional `TimerOutput` for
   recording the time spent on different functions.
+
+# Extended help
+
+## Adding an external velocity
+
+One can set the **`external_velocity`** keyword argument to impose an external velocity field
+``\bm{v}_{\text{f}}(\bm{x}, t)``.
+In this case, the total velocity at a point ``\bm{x}`` (which can be on a vortex filament) will be
+
+```math
+\bm{v}(\bm{x}) = \bm{v}_{\text{BS}}(\bm{x}) + \bm{v}_{\text{f}}(\bm{x})
+```
+
+where ``\bm{v}_{\text{BS}}`` is the velocity obtained from the Biot–Savart law.
+
+This can be a way of applying an external forcing (and injecting energy) to the vortex system.
+
+The external velocity should be given as a function, which should have the signature `v_ext(x⃗::Vec3, t::Real) -> Vec3`.
+The function should return a `Vec3` with the velocity at the location `x⃗` and time `t`.
+For example, to impose a constant sinusoidal forcing in the ``x`` direction which varies along ``z``,
+the `external_velocity` keyword argument could look like:
+
+    external_velocity = (x⃗, t) -> Vec3(0.1 * sin(2 * x⃗.z), 0, 0)
+
+One usually wants the external velocity to be divergence-free, to preserve the
+incompressibility of the flow.
+
+### Notes on energy estimation
+
+When setting an external velocity ``\bm{v}_{\text{f}}(\bm{x}, t)``, one may also want to
+impose an external streamfunction field ``\bm{ψ}_{\text{f}}(\bm{x}, t)`` via the **`external_streamfunction`**
+keyword argument.
+Such a field plays no role in the dynamics, but it allows to include the kinetic energy
+associated to the external velocity field when calling [`Diagnostics.kinetic_energy_from_streamfunction`](@ref).
+If provided, the external streamfunction should satisfy ``\bm{v}_{\text{f}} = \bm{∇} × \bm{ψ}_{\text{f}}``.
+
+More precisely, when applying an external velocity field, the total kinetic energy is
+
+```math
+\begin{align*}
+    E &= \frac{1}{2V} ∫ |\bm{v}_{\text{BS}} + \bm{v}_{\text{f}}|^2 \, \mathrm{d}^3\bm{x}
+    \\
+    &= \frac{1}{2V} \left[
+    Γ ∮_{\mathcal{C}} \left( \bm{ψ}_{\text{BS}} + 2 \bm{ψ}_{\text{f}} \right) ⋅ \mathrm{d}\bm{s}
+    +
+    ∫ |\bm{v}_{\text{f}}|^2 \, \mathrm{d}^3\bm{x}
+    \right],
+\end{align*}
+```
+
+where ``\bm{ψ}_{\text{BS}}`` is the streamfunction obtained from the Biot-Savart law.
+Accordingly, when passing `external_streamfunction`, the streamfunction values attached to
+vortex filaments (in `iter.ψs`) will actually contain
+``\bm{ψ}_{\text{BS}} + 2 \bm{ψ}_{\text{f}}`` (note the factor 2) to enable the
+computation of the full kinetic energy.
+Moreover, calling [`Diagnostics.kinetic_energy_from_streamfunction`](@ref) with a
+[`VortexFilamentSolver`](@ref) will automatically include an estimation of the kinetic
+energy of the external velocity field (the volume integral above).
+
+!!! note
+
+    Whether this total kinetic energy is of any physical interest is a different question,
+    so it can be reasonable to ignore the `external_streamfunction` parameter in order to
+    obtain the energy associated to ``\bm{v}_{\text{BS}}`` only.
 """
 function init(
         prob::VortexFilamentProblem{T}, scheme::TemporalScheme;
@@ -261,8 +331,15 @@ function init(
         reconnect::ReconnectionCriterion = NoReconnections(),
         adaptivity::AdaptivityCriterion = NoAdaptivity(),
         callback::F = identity,
+        external_velocity::ExtVel = nothing,
+        external_streamfunction::ExtStf = nothing,
         timer = TimerOutput("VortexFilament"),
-    ) where {F <: Function, T}
+    ) where {
+        F <: Function,
+        ExtVel <: Union{Nothing, Function},
+        ExtStf <: Union{Nothing, Function},
+        T,
+    }
     (; tspan,) = prob
     (; Ls,) = prob.p
     fs = convert(VectorOfVectors, prob.fs)
@@ -275,9 +352,9 @@ function init(
     cache_timestepper = init_cache(scheme, fs, vs)
 
     # Wrap functions with the timer, so that timings are estimated each time the function is called.
-    advect! = timer(advect_filaments!, "advect_filaments!")
-    rhs! = timer(update_values_at_nodes!, "update_values_at_nodes!")
-    callback_ = timer(callback, "callback")
+    advect! = timer(advect_filaments!, "Advect filaments")
+    rhs! = timer(update_values_at_nodes!, "Update values at nodes")
+    callback_ = timer(callback, "Callback")
 
     if adaptivity !== NoAdaptivity() && !can_change_dt(scheme)
         throw(ArgumentError(lazy"temporal scheme $scheme doesn't support adaptibility; set `adaptivity = NoAdaptivity()` or choose a different scheme"))
@@ -285,15 +362,55 @@ function init(
 
     time = TimeInfo(nstep = 0, t = first(tspan), dt = dt, dt_prev = dt)
 
+    external_forcing = (
+        velocity = external_velocity,
+        streamfunction = external_streamfunction,
+    )
+    check_external_streamfunction(external_forcing, Ls)
+
     iter = VortexFilamentSolver(
         prob, fs_sol, vs, ψs, time, T(dtmin), refinement, adaptivity, cache_reconnect,
-        cache_bs, cache_timestepper, fast_term, callback_, timer,
-        advect!, rhs!,
+        cache_bs, cache_timestepper, fast_term, callback_, external_forcing,
+        timer, advect!, rhs!,
     )
 
     finalise_step!(iter)
 
     iter
+end
+
+# Returns gradient of vector function f: ℝ³ ↦ ℝ³
+function vector_gradient(f::F, x⃗::Vec3) where {F <: Function}
+    ∇s = ntuple(Val(3)) do i
+        # Gradient of i-th vector component.
+        ForwardDiff.gradient(x -> f(x)[i], x⃗)
+    end
+    hcat(∇s...)  # gradient tensor (Aᵢⱼ = ∂ᵢfⱼ)
+end
+
+function curl(f::F, x⃗::Vec3) where {F <: Function}
+    A = vector_gradient(f, x⃗)
+    oftype(x⃗, (A[2, 3] - A[3, 2], A[3, 1] - A[1, 3], A[1, 2] - A[2, 1]))
+end
+
+# Check that the external streamfunction satisfies v = ∇ × ψ on a single arbitrary point.
+# We check this using automatic differentiation of ψ at a given point.
+function check_external_streamfunction(forcing::NamedTuple, Ls)
+    (; velocity, streamfunction,) = forcing
+    (velocity === nothing || streamfunction === nothing) && return nothing
+    N = length(Ls)
+    x⃗ = Vec3(ntuple(i -> 0.07 * i * Ls[i], Val(N)))  # arbitrary point in [0, L]³
+    t = 0.0  # arbitrary time
+    v⃗_actual = velocity(x⃗, t)
+    v⃗_from_streamfunction = curl(y⃗ -> streamfunction(y⃗, t), x⃗)
+    isapprox(v⃗_actual, v⃗_from_streamfunction) || @warn(
+        """
+        it seems like the external streamfunction doesn't match the external velocity.
+        This may cause wrong energy estimates.
+        """,
+        v⃗_actual, v⃗_from_streamfunction,
+    )
+    nothing
 end
 
 function _update_values_at_nodes!(
@@ -358,12 +475,35 @@ function _update_values_at_nodes!(
 end
 
 # This is the most general variant which should be called by timesteppers.
-# For now we don't use the time, but we might in the future...
 function update_values_at_nodes!(
         fields::NamedTuple, fs, t::Real, iter;
         component = Val(:full),  # compute slow + fast components by default
     )
     _update_values_at_nodes!(component, iter.fast_term, fields, fs, iter)
+    _add_external_fields!(fields, iter.external_forcing, fs, t)
+end
+
+function _add_external_fields!(fields::NamedTuple, forcing::NamedTuple, fs, t)
+    if haskey(fields, :velocity)
+        _add_external_field!(fields.velocity, forcing.velocity, fs, t)
+    end
+    if haskey(fields, :streamfunction)
+        # We multiply the external streamfunction by 2 to get the right kinetic energy.
+        _add_external_field!(fields.streamfunction, forcing.streamfunction, fs, t; factor = 2)
+    end
+    fields
+end
+
+_add_external_field!(vs_all, ::Nothing, args...; kws...) = vs_all  # do nothing
+
+function _add_external_field!(vs_all, vext::F, fs, time; factor = 1) where {F <: Function}
+    @assert eachindex(vs_all) == eachindex(fs)
+    for (f, vs) ∈ zip(fs, vs_all)
+        for i ∈ eachindex(vs, f)
+            @inbounds vs[i] = vs[i] + factor * vext(f[i], time)
+        end
+    end
+    vs_all
 end
 
 # Case where only the velocity is passed (generally used in internal RK substeps).